+++ HISTORY.md
@@ -49,6 +49,7 @@ Items marked [*] are breaking changes.
 - `file` and `directory` doers now error with a meaningful message if you try
   to ensure a path which already exists, but is of a different type.
 - `smf` doer now lets you set service properties.
+- Fix bug where removing a service property always produced an error.
 
 ### Internals